refactor: update Spinner import paths for consistency across components
